add support for fs-uae amiga
* amiga first

* amiga next

* amiga dev

* fixes, refactoring and manage img files

big update

+++ src/main/java/com/notamethod/fluppy/core/game/GameManager.java
@@ -4,13 +4,17 @@ import com.notamethod.fluppy.core.ApplicationDatabase;
 import com.notamethod.fluppy.core.preferences.PreferencesBean;
 import com.notamethod.fluppy.dosbox.DosBoxException;
 import com.notamethod.fluppy.dosbox.DosBoxManager;
+import com.notamethod.fluppy.dosbox.UAEManager;
 import com.notamethod.fluppy.gui.PanelListener;
 import com.notamethod.fluppy.util.HelperClass;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
 
 @Slf4j
@@ -19,6 +23,7 @@ public class GameManager {
     private ApplicationDatabase applicationDatabase;
     private PreferencesBean preferences;
     private DosBoxManager dosBoxManager;
+    private UAEManager uaeManager;
     public GameManager(ApplicationDatabase applicationDatabase, PreferencesBean preferences, DosBoxManager dosBoxManager) {
         this.applicationDatabase = applicationDatabase;
         this.preferences=preferences;
@@ -28,7 +33,7 @@ public class GameManager {
     public int deleteGame(GameApp gameApp) {
         int nbDeleted = 0;
 
-        if (gameApp.getGamePath() != null) {
+        if (gameApp.getGamePath() != null && HelperClass.gameIsInGameDir(gameApp)) {
             try {
                 HelperClass.deleteDirectory(gameApp.getGamePath());
             } catch (IOException e) {
@@ -81,6 +86,15 @@ public class GameManager {
         log.debug("adding game ->{} <- to database", game.getName());
         List<GameEntity> entiites = applicationDatabase.findGameByUnique(game.getName(), game.getYear(),game.getLanguage());
         if (!entiites.isEmpty()) {
+            if (entiites.size()==1){
+                if (game.getDiskNumber()>0){
+                    GameEntity storedGame = entiites.getFirst();
+                    if (!isDiskPresent(game,storedGame)){
+                        applicationDatabase.saveGame(storedGame);
+                        return;
+                    }
+                }
+            }
             StringBuilder b = new StringBuilder();
             for (GameEntity gamelog:entiites){
                 b.append(gamelog.getId()).append("/").append(gamelog.getGame()).append("/")
@@ -99,6 +113,41 @@ public class GameManager {
             }
         }
         save(game);
+    }
+
+    private Boolean isDiskPresent(GameApp game, GameEntity storedGame) {
+        List<String> diskList = new ArrayList<>();
+        if (storedGame.getExtraDisks()!=null){
+            String[] disks = storedGame.getExtraDisks().split(";");
+            diskList = new ArrayList<>(Arrays.asList(disks));
+        }
+
+       // Map<String,String> diskSet = new HashMap<>();
+        Boolean found = null;
+
+       if (game.getDiskNumber()>1){
+           found=false;
+           for (String disk:diskList){
+               String[] diskInfo=disk.split("#");
+             //  diskSet.put(diskInfo[0],diskInfo[1]);
+               if (String.valueOf(game.getDiskNumber()).equals(diskInfo[0])){
+                   found=true;
+                   break;
+               }
+           }
+           if (!found){
+               String newDisk = game.getDiskNumber()+"#"+game.getGamePath().toAbsolutePath().toString();
+               diskList.add(newDisk);
+               storedGame.setExtraDisks(String.join( ";",diskList));
+           }
+       }else{
+           if (storedGame.getExePath()==null || storedGame.getExePath().equals("")){
+               storedGame.setExePath(game.getExePath().toAbsolutePath().toString());
+               found= false;
+           }
+
+       }
+       return found;
     }
 
     public void updateTime(GameApp game, Long time) {
@@ -149,34 +198,71 @@ public class GameManager {
     public long runGame(GameApp game, String screenRez, PanelListener listener) throws DosBoxException {
 
         AtomicReference<Long> duration= new AtomicReference<>(0L);
-        dosBoxManager.runApplication(
-                game.getGameExe(),
-                game,
-                screenRez,
-                listener,
-                line -> log.info("[DOSBOX] "+line),
-                err -> log.error("[DOSBOX] " + err ),
-                result -> {
+        if (game.getPlatform()!=null && game.getPlatform().contains("amiga")){
+            if (uaeManager==null){
+                uaeManager = new UAEManager();
+            }
+            uaeManager.runApplication(
+                    game.getGameExe(),
+                    game,
+                    screenRez,
+                    listener,
+                    line -> log.info("[FSUAE] " + line),
+                    err -> log.error("[FSUAE] " + err),
+                    result -> {
 
-                    if (listener != null) {
+                        if (listener != null) {
 
-                        listener.onExitGame();
-                    }
-                    if (result.success) {
-                        System.out.println("DOSBox OK");
-                    } else {
-                        System.out.println("Erreur : " + result.error);
-                    }
+                            listener.onExitGame();
+                        }
+                        if (result.success) {
+                            System.out.println("DOSBox OK");
+                        } else {
+                            System.out.println("Erreur : " + result.error);
+                        }
 
-                    System.out.println("Durée : " + result.durationMillis + " ms");
-                    duration.set(result.durationMillis);
-                    if (duration.get()>0){
-                        updateTime(game, duration.get()/1000);
+                        System.out.println("Durée : " + result.durationMillis + " ms");
+                        duration.set(result.durationMillis);
+                        if (duration.get() > 0) {
+                            updateTime(game, duration.get() / 1000);
+                        }
+                        System.out.println("Exit code : " + result.exitCode);
                     }
-                    System.out.println("Exit code : " + result.exitCode);
-                }
-        );
-        return duration.get();
+            );
+
+            return duration.get();
+
+        }else {
+            dosBoxManager.runApplication(
+                    game.getGameExe(),
+                    game,
+                    screenRez,
+                    listener,
+                    line -> log.info("[DOSBOX] " + line),
+                    err -> log.error("[DOSBOX] " + err),
+                    result -> {
+
+                        if (listener != null) {
+
+                            listener.onExitGame();
+                        }
+                        if (result.success) {
+                            System.out.println("DOSBox OK");
+                        } else {
+                            System.out.println("Erreur : " + result.error);
+                        }
+
+                        System.out.println("Durée : " + result.durationMillis + " ms");
+                        duration.set(result.durationMillis);
+                        if (duration.get() > 0) {
+                            updateTime(game, duration.get() / 1000);
+                        }
+                        System.out.println("Exit code : " + result.exitCode);
+                    }
+            );
+
+            return duration.get();
+        }
 
     }
 
@@ -201,5 +287,9 @@ public class GameManager {
 
     public void save(CompanyEntity company) {
          applicationDatabase.saveCompany(company);
+    }
+
+    public DosBoxManager getDosBoxManager() {
+        return dosBoxManager;
     }
 }
